Split network module into multiple sub-modules
