+++ cnd/Cargo.toml
@@ -6,7 +6,6 @@ edition = "2018"
 description = "Reference implementation of a COMIT network daemon."
 
 [dependencies]
-ambassador = "0.2"
 anyhow = "1"
 async-trait = "0.1"
 atty = "0.2"
