+++ cnd/src/http_api/peers.rs
@@ -1,4 +1,4 @@
-use crate::{http_api::Http, network::ComitPeers, Facade};
+use crate::{http_api::Http, Facade};
 use libp2p::{Multiaddr, PeerId};
 use serde::Serialize;
 use warp::{Rejection, Reply};
@@ -6,7 +6,8 @@ use warp::{Rejection, Reply};
 #[allow(clippy::needless_pass_by_value)]
 pub async fn get_peers(facade: Facade) -> Result<impl Reply, Rejection> {
     let peers = facade
-        .comit_peers()
+        .swarm
+        .connected_peers()
         .await
         .map(|(peer, addresses)| Peer {
             id: Http(peer),
